setup tests and reorganize testable functions
