+++ tests/sunburst.helper.test.ts
@@ -1,0 +1,7 @@
+describe('PluginSunburst Helper tests', () => {
+    describe('PlotlyPayload to CellSetSelection', () => {
+        test('First empty test', () => {
+            expect(true).toBe(true);
+        });
+    });
+});
